requirements.txt
py-cord[voice]==2.6.1
google-genai>=0.3.0
python-dotenv>=1.0.0
PyNaCl>=1.5.0
